requirements.txt
Django==4.2.7
mysqlclient==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
stripe==7.0.0
django-cors-headers==4.3.0
python-decouple==3.8
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.29.0
